Support for directories and page crumbs
This changeset adds support for displaying a directory as an index()
of the files inside it, and also adds "page crumbs" (links to each
component of the path) to each page and directory template.

Also, added a 'SORT' option in config.py; set it to 'title' or
'url' to control the order in which index() sorts pages.