Replacing the page list on directory pages with a tree
This changeset will render the list of pages and subdirectories
of a directory as a tree of nested lists, instead of a table.

Also, this changeset changes the fuctionality of the "home" page.
If there exists a page tagged 'HOMEPAGE', that's used for the '/'
url; otherwise, the full tree of lists is displayed.